+++ server/prisma/migrations/20260925134347_modalidad_pago_otro/migration.sql
@@ -1,0 +1,5 @@
+-- AlterEnum
+ALTER TYPE "ModalidadPago" ADD VALUE 'OTRO';
+
+-- AlterTable
+ALTER TABLE "projects" ADD COLUMN     "modalidadPagoNota" TEXT;